+++ packages/modules/imodel-browser/src/mui/components/baseCard/BaseCardThumbnailArea.tsx
