+++ packages/modules/imodel-browser/src/mui/components/baseCard/SvgThumbnail.tsx
