+++ packages/modules/imodel-browser/src/mui/components/baseCard/ThumbnailIconButton.tsx
