+++ packages/modules/imodel-browser/src/mui/components/tileFavoriteIcon/FavoriteIconMUI.tsx
@@ -64,7 +64,9 @@ export const FavoriteIconMUI = ({
       onMouseLeave={() => setHovered(false)}
       onClick={async (event: React.MouseEvent<HTMLButtonElement>) => {
         // debounce
-        if (pending) return;
+        if (pending) {
+          return;
+        }
 
         setPending(true);
         try {
