+++ packages/modules/imodel-browser/src/mui/containers/ITwinGrid/ITwinTileMUI.tsx
@@ -6,20 +6,20 @@ import Chip from "@mui/material/Chip";
 import svgItwin from "@stratakit/icons/itwin.svg";
 import React from "react";
 
+import { ITwinTileProps } from "../../../containers/ITwinGrid/ITwinTile";
+import { ITwinFull } from "../../../types";
+import { _mergeStrings } from "../../../utils/_apiOverrides";
+import {
+  type MoreActionsMenuItemMUI,
+  type ResolvedMoreActionsMenuItem,
+} from "../../../utils/_buildMenuOptions";
+import { formatDate } from "../../../utils/formatDate";
 import {
   type BaseCardProps,
   BaseCard,
 } from "../../components/baseCard/BaseCard";
 import { SvgThumbnail } from "../../components/baseCard/SvgThumbnail";
 import { FavoriteIconMUI } from "../../components/tileFavoriteIcon/FavoriteIconMUI";
-import { ITwinFull } from "../../types";
-import { _mergeStrings } from "../../utils/_apiOverrides";
-import {
-  type MoreActionsMenuItemMUI,
-  type ResolvedMoreActionsMenuItem,
-} from "../../utils/_buildMenuOptions";
-import { formatDate } from "../../utils/formatDate";
-import { ITwinTileProps } from "./ITwinTile";
 
 /** @alpha */
 export interface ITwinTilePropsMUI
