+++ packages/modules/imodel-browser/src/mui/containers/iModelGrid/IModelGridMUI.tsx
@@ -6,35 +6,38 @@ import Box from "@mui/material/Box";
 import React from "react";
 import { InView } from "react-intersection-observer";
 
-import { BaseCardLoading } from "../../components/baseCard/BaseCardLoading";
-import { NoResultsMUI as NoResults } from "../../components/noResults/NoResultsMUI";
-import { IModelFavoritesProvider } from "../../contexts/IModelFavoritesContext";
-import { type IModelTableOverridesMUI } from "../../mui/types";
+import type { IModelGridProps } from "../../../containers/iModelGrid/IModelGrid";
+import {
+  DEFAULT_PAGE_SIZE,
+  useIModelData,
+} from "../../../containers/iModelGrid/useIModelData";
+import { IModelFavoritesProvider } from "../../../contexts/IModelFavoritesContext";
 import {
   type AccessTokenProvider,
   type ApiOverrides,
   type IModelFull,
   DataStatus,
   IModelSortOptions,
-} from "../../types";
-import { _mergeStrings } from "../../utils/_apiOverrides";
+} from "../../../types";
+import { _mergeStrings } from "../../../utils/_apiOverrides";
 import {
   type CardActionsItemMUI,
   MoreActionsMenuItemMUI,
   resolveCardActionsItemsMUI,
-} from "../../utils/_buildMenuOptions";
+} from "../../../utils/_buildMenuOptions";
 import {
   addIModelToRecents,
   removeIModelFromRecents,
-} from "../../utils/iModelApi";
+} from "../../../utils/iModelApi";
+import { BaseCardLoading } from "../../components/baseCard/BaseCardLoading";
+import { NoResultsMUI as NoResults } from "../../components/noResults/NoResultsMUI";
+import { type IModelTableOverridesMUI } from "../../types";
 import {
   type IModelTileMUIProps,
   IModelTileMUI,
 } from "../iModelTiles/IModelTileMUI";
 import { clientSideIModelSort } from "./clientSideIModelSort";
-import type { IModelGridProps } from "./IModelGrid";
 import { type IModelTableMUIStrings, IModelTableMUI } from "./IModelTableMUI";
-import { DEFAULT_PAGE_SIZE, useIModelData } from "./useIModelData";
 
 /**
  * Localized strings for the MUI IModelGrid. Extends the table-level strings
